integrate CaptureValueChangeTransformer into BackInTimeDebuggableMethodInvocationTransformer
